Add files via upload
The SDLC Model Selection Dataset provides comprehensive data on various factors that influence the selection of Software Development Life Cycle (SDLC) models for software projects. This dataset is designed to support research and academic projects that involve the analysis, comparison, and prediction of SDLC models based on project-specific characteristics and requirements.